+++ virtualizarr/zarr.py
@@ -5,6 +5,7 @@ from typing import (
     Any,
     Literal,
     NewType,
+    Optional,
 )
 
 import numpy as np
@@ -75,7 +76,7 @@ class ZArray(BaseModel):
     def from_kerchunk_refs(cls, decoded_arr_refs_zarray) -> "ZArray":
         # coerce type of fill_value as kerchunk can be inconsistent with this
         fill_value = decoded_arr_refs_zarray["fill_value"]
-        if fill_value is None or fill_value == "NaN":
+        if fill_value is None or fill_value == "NaN" or fill_value == "nan":
             fill_value = np.nan
 
         return ZArray(
@@ -101,6 +102,31 @@ class ZArray(BaseModel):
 
     def to_kerchunk_json(self) -> str:
         return ujson.dumps(self.dict())
+
+    def replace(
+        self,
+        chunks: Optional[tuple[int, ...]] = None,
+        compressor: Optional[str] = None,
+        dtype: Optional[np.dtype] = None,
+        fill_value: Optional[float] = None,  # float or int?
+        filters: Optional[list[dict]] = None,  # type: ignore[valid-type]
+        order: Optional[Literal["C"] | Literal["F"]] = None,
+        shape: Optional[tuple[int, ...]] = None,
+        zarr_format: Optional[Literal[2] | Literal[3]] = None,
+    ) -> "ZArray":
+        """
+        Convenience method to create a new ZArray from an existing one by altering only certain attributes.
+        """
+        return ZArray(
+            chunks=chunks if chunks is not None else self.chunks,
+            compressor=compressor if compressor is not None else self.compressor,
+            dtype=dtype if dtype is not None else self.dtype,
+            fill_value=fill_value if fill_value is not None else self.fill_value,
+            filters=filters if filters is not None else self.filters,
+            shape=shape if shape is not None else self.shape,
+            order=order if order is not None else self.order,
+            zarr_format=zarr_format if zarr_format is not None else self.zarr_format,
+        )
 
 
 def encode_dtype(dtype: np.dtype) -> str: